Fully connected neural network from Matlab executed in C
Export a fully connected neural network from Matlab to be executed in C
using the weights. Valid for any fully connected architecture (layers
and weights). Activation function by default is tansig for all but the
last layer which is purelin. Allows for transformation of input and back
transformation of output automatically (scaling and centering).